requirements.txt
albumentations==0.5.2
gym==0.17.*
dtw>=1.3.3
pymunk==5.6.0
ml-collections==0.1.0
scikit-learn==0.24.1
tensorflow-cpu==2.6.0
torch==1.7.1
imageio
imageio-ffmpeg
tqdm
torchvision==0.8.2
absl-py
numpy
scipy
gdown==3.13.1
PyPDF2
git+https://github.com/kevinzakka/torchkit.git@v0.0.2
git+https://github.com/kevinzakka/x-magical.git@v0.0.2
